requirements.txt
# Core AI
anthropic>=0.40.0
google-genai>=0.1.0

# Data Science
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0

# PDF Generation
fpdf2>=2.7.0

# UI
streamlit>=1.35.0

# Utilities
python-dotenv>=1.0.0
